refactor: split limit hypothesis evaluators
